+++ ChangeLog.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [0.1.2.0] - 2019-12-5
+### Added
+- wrike-426952824 : added Ecoli and CHO codon tables
+- forbidden sequences as parameter
+
 ## [0.1.1.0] - 2019-11-13
 ### Added
 - Codon optimization tool